requirements.txt
streamlit
transformers
torch
sentencepiece
huggingface_hub
sacrebleugit commit -m "initial commit"